// src/zaken-generated-client/1.4.1/api/data-contracts.ts
export interface AuditTrail {
  /**
   * uuid
   * Unieke identificatie van de audit regel.
   * @format uuid
   */
  uuid?: string;
  /**
   * bron
   * De naam van het component waar de wijziging in is gedaan.
   *
   * Uitleg bij mogelijke waarden:
   *
   * * `ac` - Autorisaties API
   * * `nrc` - Notificaties API
   * * `zrc` - Zaken API
   * * `ztc` - Catalogi API
   * * `drc` - Documenten API
   * * `brc` - Besluiten API
   * * `cmc` - Contactmomenten API
   * * `kc` - Klanten API
   * * `vrc` - Verzoeken API
   */
  bron: BronEnum;
  /**
   * applicatie id
   * Unieke identificatie van de applicatie, binnen de organisatie.
   * @maxLength 100
   */
  applicatieId?: string;
  /**
   * applicatie weergave
   * Vriendelijke naam van de applicatie.
   * @maxLength 200
   */
  applicatieWeergave?: string;
  /**
   * gebruikers id
   * Unieke identificatie van de gebruiker die binnen de organisatie herleid kan worden naar een persoon.
   * @maxLength 255
   */
  gebruikersId?: string;
  /**
   * gebruikers weergave
   * Vriendelijke naam van de gebruiker.
   * @maxLength 255
   */
  gebruikersWeergave?: string;
  /**
   * actie
   * De uitgevoerde handeling.
   *
   * De bekende waardes voor dit veld zijn hieronder aangegeven,                         maar andere waardes zijn ook toegestaan
   *
   * Uitleg bij mogelijke waarden:
   *
   * * `create` - Object aangemaakt
   * * `list` - Lijst van objecten opgehaald
   * * `retrieve` - Object opgehaald
   * * `destroy` - Object verwijderd
   * * `update` - Object bijgewerkt
   * * `partial_update` - Object deels bijgewerkt
   * @maxLength 50
   */
  actie: string;
  /**
   * actie weergave
   * Vriendelijke naam van de actie.
   * @maxLength 200
   */
  actieWeergave?: string;
  /**
   * resultaat
   * HTTP status code van de API response van de uitgevoerde handeling.
   * @min 100
   * @max 599
   */
  resultaat: number;
  /**
   * hoofd object
   * De URL naar het hoofdobject van een component.
   * @format uri
   * @maxLength 1000
   */
  hoofdObject: string;
  /**
   * resource
   * Het type resource waarop de actie gebeurde.
   * @maxLength 50
   */
  resource: string;
  /**
   * resource url
   * De URL naar het object.
   * @format uri
   * @maxLength 1000
   */
  resourceUrl: string;
  /**
   * toelichting
   * Toelichting waarom de handeling is uitgevoerd.
   */
  toelichting?: string;
  /**
   * resource weergave
   * Vriendelijke identificatie van het object.
   * @maxLength 200
   */
  resourceWeergave: string;
  /**
   * aanmaakdatum
   * De datum waarop de handeling is gedaan.
   * @format date-time
   */
  aanmaakdatum: string;
  /** wijzigingen */
  wijzigingen: Wijzigingen;
}
// ...
export enum BronEnum {
  Ac = "ac",
  Nrc = "nrc",
  Zrc = "zrc",
  Ztc = "ztc",
  Drc = "drc",
  Brc = "brc",
  Cmc = "cmc",
  Kc = "kc",
  Vrc = "vrc",
}
// ...
export interface FieldValidationError {
  /**
   * name
   * Naam van het veld met ongeldige gegevens
   */
  name: string;
  /**
   * code
   * Systeemcode die het type fout aangeeft
   */
  code: string;
  /**
   * reason
   * Uitleg wat er precies fout is met de gegevens
   */
  reason: string;
}

/** Formaat van HTTP 4xx en 5xx fouten. */
export interface Fout {
  /**
   * type
   * URI referentie naar het type fout, bedoeld voor developers
   */
  type?: string;
  /**
   * code
   * Systeemcode die het type fout aangeeft
   */
  code: string;
  /**
   * title
   * Generieke titel voor het type fout
   */
  title: string;
  /**
   * status
   * De HTTP status code
   */
  status: number;
  /**
   * detail
   * Extra informatie bij de fout, indien beschikbaar
   */
  detail: string;
  /**
   * instance
   * URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.
   */
  instance: string;
}
// ...
export interface PaginatedZaakList {
  /** @example 123 */
  count?: number;
  /**
   * @format uri
   * @example "http://api.example.org/accounts/?page=4"
   */
  next?: string | null;
  /**
   * @format uri
   * @example "http://api.example.org/accounts/?page=2"
   */
  previous?: string | null;
  results?: Zaak[];
}
// ...
export interface PatchedZaakEigenschap {
  /**
   * url
   * @format uri
   */
  url?: string;
  /**
   * uuid
   * Unieke resource identifier (UUID4)
   * @format uuid
   */
  uuid?: string;
  /**
   * zaak
   * @format uri
   */
  zaak?: string;
  /**
   * eigenschap
   * URL-referentie naar de EIGENSCHAP (in de Catalogi API).
   * @format uri
   * @maxLength 1000
   */
  eigenschap?: string;
  /**
   *  naam
   * De naam van de EIGENSCHAP (overgenomen uit de Catalogi API).
   */
  naam?: string;
  /** waarde */
  waarde?: string;
}
// ...
export interface ValidatieFout {
  /**
   * type
   * URI referentie naar het type fout, bedoeld voor developers
   */
  type?: string;
  /**
   * code
   * Systeemcode die het type fout aangeeft
   */
  code: string;
  /**
   * title
   * Generieke titel voor het type fout
   */
  title: string;
  /**
   * status
   * De HTTP status code
   */
  status: number;
  /**
   * detail
   * Extra informatie bij de fout, indien beschikbaar
   */
  detail: string;
  /**
   * instance
   * URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.
   */
  instance: string;
  /** invalidParams */
  invalidParams: FieldValidationError[];
}
// ...
export interface Wijzigingen {
  /**
   * oud
   * Volledige JSON body van het object zoals dat bestond voordat de actie heeft plaatsgevonden.
   */
  oud?: Record<string, any>;
  /**
   * nieuw
   * Volledige JSON body van het object na de actie.
   */
  nieuw?: Record<string, any>;
}
// ...
export interface ZaakBesluit {
  /**
   * url
   * @format uri
   */
  url: string;
  /**
   * uuid
   * Unieke resource identifier (UUID4)
   * @format uuid
   */
  uuid: string;
  /**
   * besluit
   * URL-referentie naar het BESLUIT (in de Besluiten API), waar ook de relatieinformatie opgevraagd kan worden.
   * @format uri
   * @maxLength 1000
   */
  besluit: string;
}
// ...
export interface ZaakEigenschap {
  /**
   * url
   * @format uri
   */
  url: string;
  /**
   * uuid
   * Unieke resource identifier (UUID4)
   * @format uuid
   */
  uuid: string;
  /**
   * zaak
   * @format uri
   */
  zaak: string;
  /**
   * eigenschap
   * URL-referentie naar de EIGENSCHAP (in de Catalogi API).
   * @format uri
   * @maxLength 1000
   */
  eigenschap: string;
  /**
   *  naam
   * De naam van de EIGENSCHAP (overgenomen uit de Catalogi API).
   */
  naam: string;
  /** waarde */
  waarde: string;
}
// ...
export type ZaakListData = PaginatedZaakList;

export type ZaakCreateData = Zaak;

export type ZaakRetrieveData = Zaak;

export type ZaakUpdateData = Zaak;

export type ZaakPartialUpdateData = Zaak;

export type ZaakDestroyData = any;

export type ZaakHeadersData = any;

export type AudittrailListData = AuditTrail[];

export type AudittrailRetrieveData = AuditTrail;

export type ZaakbesluitListData = ZaakBesluit[];

export type ZaakbesluitCreateData = ZaakBesluit;

export type ZaakbesluitRetrieveData = ZaakBesluit;

export type ZaakbesluitDestroyData = any;

export type ZaakeigenschapListData = ZaakEigenschap[];

export type ZaakeigenschapCreateData = ZaakEigenschap;

export type ZaakeigenschapRetrieveData = ZaakEigenschap;

export type ZaakeigenschapUpdateData = ZaakEigenschap;

export type ZaakeigenschapPartialUpdateData = ZaakEigenschap;

export type ZaakeigenschapDestroyData = any;

export type ZaakeigenschapHeadersData = any;

export interface ZaakZoekParams {
  /** Een pagina binnen de gepagineerde set resultaten. */
  page?: number;
}

export type ZaakZoekData = PaginatedZaakList;